modify some names
